Move Javascript into assets folder
This adds to the work to separate the client code from the server
code.
