+++ assets/javascripts/_deprecated/company-add.js
@@ -1,4 +1,4 @@
-const { addClass, removeClass } = require('../../lib/element-stuff')
+const { addClass, removeClass } = require('./lib/element-stuff')
 
 const CompanyAdd = {
   init () {
