+++ assets/javascripts/_deprecated/lib/element-stuff.js
