+++ common/transform-sectors.js
