+++ test/unit/javascripts/sortable-table.test.js
@@ -1,6 +1,6 @@
 const jsdom = require('jsdom')
 const { JSDOM } = jsdom
-const SortableTable = require(`~/src/javascripts/modules/sortable-table`)
+const SortableTable = require(`~/assets/javascripts/modules/sortable-table`)
 const { domTokenToArray } = require('~/test/unit/component-helper')
 
 const FIRST_NAME_SELECTOR = '.js-SortableTable tbody tr:first-child td:first-child'
